refactor(optimizer): s_expr.rs
